election test data: ASCII-scheme instead of JSON
